src/chiaro/core.py
from pydantic import BaseModel


class DynamicModel(BaseModel):
    """Base class for dynamically created Pydantic models."""

    @classmethod
    def from_schema(cls, schema: dict) -> type["DynamicModel"]:
        """
        Create a dynamic model class from a given JSON schema.

        :param schema: The JSON schema to create the model from.
        :return: A dynamically created Pydantic model class.
        """
        # Logic to dynamically create a model from a schema

    def to_schema(self) -> dict:
        """
        Convert the model instance back to a JSON schema.

        :return: The JSON schema representing the model.
        """
        # Logic to convert the model instance back to a JSON schema


class SchemaParser:
    """Class to parse JSON schemas and create dynamic Pydantic models."""

    def __init__(self, schema: dict):
        self.schema = schema

    def parse(self) -> type[DynamicModel]:
        """
        Parse the schema and create a dynamic model.

        :return: A dynamically created Pydantic model class.
        """
        # Logic to parse the schema and create a dynamic model


class SchemaGenerator:
    """Class to generate JSON schemas from Pydantic models."""

    def __init__(self, model: type[DynamicModel]):
        self.model = model

    def generate(self) -> dict:
        """
        Generate a JSON schema from the given model.

        :return: The generated JSON schema.
        """
        # Logic to generate a JSON schema from the given model
